+++ ios/RCTConvert+BitmovinPlayer.swift
@@ -42,6 +42,9 @@ extension RCTConvert {
         if let networkConfig = RCTConvert.networkConfig(json["networkConfig"]) {
             playerConfig.networkConfig = networkConfig
         }
+        if let nowPlayingConfig = RCTConvert.mediaControlConfig(json["mediaControlConfig"]) {
+            playerConfig.nowPlayingConfig = nowPlayingConfig
+        }
 #if os(iOS)
         if let remoteControlConfig = RCTConvert.remoteControlConfig(json["remoteControlConfig"]) {
             playerConfig.remoteControlConfig = remoteControlConfig
@@ -1239,9 +1242,13 @@ extension RCTConvert {
         guard let json = json as? [String: Any?] else {
             return nil
         }
+        let variant = json["variant"] as? [String: Any?]
+        let uiManagerFactoryFunction = variant?["uiManagerFactoryFunction"] as? String
+        let defaultUiManagerFactoryFunction = "bitmovin.playerui.UIFactory.buildDefaultSmallScreenUI"
 
         return RNUiConfig(
-            playbackSpeedSelectionEnabled: json["playbackSpeedSelectionEnabled"] as? Bool ?? true
+            playbackSpeedSelectionEnabled: json["playbackSpeedSelectionEnabled"] as? Bool ?? true,
+            uiManagerFactoryFunction: uiManagerFactoryFunction ?? defaultUiManagerFactoryFunction
         )
     }
 
@@ -1328,6 +1335,18 @@ extension RCTConvert {
             "body": toJson(data: httpResponse.body)
         ]
     }
+
+    static func mediaControlConfig(_ json: Any?) -> NowPlayingConfig? {
+        let nowPlayingConfig = NowPlayingConfig()
+        guard let json = json as? [String: Any?] else {
+            nowPlayingConfig.isNowPlayingInfoEnabled = true
+            return nowPlayingConfig
+        }
+        if let isEnabled = json["isEnabled"] as? Bool {
+            nowPlayingConfig.isNowPlayingInfoEnabled = isEnabled
+        }
+        return nowPlayingConfig
+    }
 }
 /**
  * React native specific PlayerViewConfig.
@@ -1369,6 +1388,7 @@ internal struct RNPlayerViewConfig {
  */
 internal struct RNUiConfig {
     let playbackSpeedSelectionEnabled: Bool
+    let uiManagerFactoryFunction: String
 }
 
 /**